refactor(alias): reorganize file structure and refine documentation
