+++ alias/src/alias/runtime/runtime_compat/adapter/alias_stream_adapter.py
